+++ src/libcsg/map.cc
@@ -121,7 +121,7 @@ void Map_Sphere::Initialize(const Molecule *in, Bead *out, Property *opts_bead,
   double norm = 1. / std::accumulate(weights.begin(), weights.end(), 0.);
 
   transform(weights.begin(), weights.end(), weights.begin(),
-            bind2nd(multiplies<double>(), norm));
+            std::bind2nd(multiplies<double>(), norm));
   // get the d vector if exists or initialize same as weights
   vector<double> d;
   if (_opts_map->exists("d")) {
@@ -130,7 +130,7 @@ void Map_Sphere::Initialize(const Molecule *in, Bead *out, Property *opts_bead,
     // normalize d coefficients
     norm = 1. / std::accumulate(d.begin(), d.end(), 0.);
     transform(d.begin(), d.end(), d.begin(),
-              bind2nd(multiplies<double>(), norm));
+              std::bind2nd(multiplies<double>(), norm));
   } else {
     // initialize force-weights with weights
     d.resize(weights.size());